+++ src/Catalog/AbstractCatalogServiceDecorator.php
@@ -12,6 +12,7 @@ namespace Wizaplace\SDK\Catalog;
 use Psr\Http\Message\ResponseInterface;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
+use Wizaplace\SDK\SortDirection;
 
 /**
  * Class AbstractCatalogServiceDecorator
@@ -98,11 +99,14 @@ abstract class AbstractCatalogServiceDecorator implements CatalogServiceInterfac
     }
 
     /**
+     * @param string $criteria
+     * @param string $direction
+     *
      * @return CategoryTree[]
      */
-    public function getCategoryTree(): array
+    public function getCategoryTree(string $criteria = CategorySortCriteria::POSITION, string $direction = SortDirection::ASC): array
     {
-        return $this->decorated->getCategoryTree();
+        return $this->decorated->getCategoryTree($criteria, $direction);
     }
 
     /**